+++ src/communication/azext_communication/tests/latest/test_communication_rooms_scenario.py
@@ -9,6 +9,7 @@ from azure.cli.testsdk import ResourceGroupPreparer, ScenarioTest
 from azure.core.exceptions import HttpResponseError
 from .preparers import CommunicationResourcePreparer
 from .recording_processors import URIIdentityReplacer, BodyReplacerProcessor
+from datetime import datetime, timedelta
 
 class CommunicationRoomsScenarios(ScenarioTest):
 
@@ -30,7 +31,7 @@ class CommunicationRoomsScenarios(ScenarioTest):
 
         return res['properties']['id']
 
-
+    # room management test
     @ResourceGroupPreparer(name_prefix='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
     @CommunicationResourcePreparer(resource_group_parameter_name='rg')
     def test_rooms_create(self, communication_resource_info):
@@ -38,34 +39,7 @@ class CommunicationRoomsScenarios(ScenarioTest):
         room = self.cmd('az communication rooms create').get_output_in_json()
         
         id = room['id']
-        participants = room['participants']
-        roomJoinPolicy = room['roomJoinPolicy']
-        
         assert len(id) > 0
-        assert len (participants) == 0
-        assert roomJoinPolicy == 'InviteOnly'
-
-
-    @ResourceGroupPreparer(name_prefix='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
-    @CommunicationResourcePreparer(resource_group_parameter_name='rg')
-    def test_rooms_create_bad_join_policy(self, communication_resource_info):
-        os.environ['AZURE_COMMUNICATION_CONNECTION_STRING'] = communication_resource_info[1]
-
-        # add invalid join policy to the rooms
-        self.kwargs.update({
-            'join_policy': 'azure12345'})
- 
-        with self.assertRaises(HttpResponseError) as raises:
-            self.cmd('az communication rooms create --join-policy {join_policy}')
-
-        assert 'Bad Request' in str(raises.exception.reason)
-        assert raises.exception.status_code == 400
-        
-        res = raises.exception.model.additional_properties
-        assert res['status'] == 400
-        assert res['traceId'] is not None
-        assert 'validation errors' in str(res['title'])
-        
 
     @ResourceGroupPreparer(name_prefix='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
     @CommunicationResourcePreparer(resource_group_parameter_name='rg')
@@ -93,32 +67,9 @@ class CommunicationRoomsScenarios(ScenarioTest):
             self.cmd('az communication rooms delete --room {room_id} --yes')  
 
         assert 'Bad Request' in str(raises.exception.reason)
-        assert raises.exception.status_code == 400
-        
-        err = raises.exception.model.error
-        assert err.code == 'InvalidInput'
-        assert 'Invalid Room ID length' in err.message
-        
-
-    @ResourceGroupPreparer(name_prefix ='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
-    @CommunicationResourcePreparer(resource_group_parameter_name='rg')
-    def test_rooms_delete_with_invalid_id(self, communication_resource_info):
-        os.environ['AZURE_COMMUNICATION_CONNECTION_STRING'] = communication_resource_info[1]
-       
-        # delete valid room with invalid room id
-        self.kwargs.update({
-            'room_id': 'f-123456789012345'})
-        
-        with self.assertRaises(HttpResponseError) as raises:
-            self.cmd('az communication rooms delete --room {room_id} --yes')  
-
-        assert 'Bad Request' in str(raises.exception.reason)
-        assert raises.exception.status_code == 400
-        
-        err = raises.exception.model.error
-        assert err.code == 'InvalidInput'
-        assert 'Room ID is invalid' in err.message
-        
+        assert raises.exception.status_code == 400       
+        assert 'Invalid room ID' in str(raises.exception)
+               
 
     @ResourceGroupPreparer(name_prefix ='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
     @CommunicationResourcePreparer(resource_group_parameter_name='rg')
@@ -134,50 +85,6 @@ class CommunicationRoomsScenarios(ScenarioTest):
 
         assert room['id'] == get_room['id']
     
-
-    @ResourceGroupPreparer(name_prefix ='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
-    @CommunicationResourcePreparer(resource_group_parameter_name='rg')
-    def test_rooms_update_join_policy(self, communication_resource_info):
-        os.environ['AZURE_COMMUNICATION_CONNECTION_STRING'] = communication_resource_info[1]
-        
-        # create a room first, it should be in the future so we can update the policy 
-        room = self.cmd('az communication rooms create --valid-from 2023-04-01T13:35').get_output_in_json()
-        
-        self.kwargs.update({
-            'room_id': room['id'],
-            'join_policy': 'CommunicationServiceUsers'})
-
-        # update the room join policy
-        updated_room = self.cmd('az communication rooms update --room {room_id} --join-policy {join_policy}').get_output_in_json()
-        
-        assert room['id'] == updated_room['id']
-        assert room['roomJoinPolicy'] == "InviteOnly"
-        assert updated_room['roomJoinPolicy'] == "CommunicationServiceUsers"
-
-
-    @ResourceGroupPreparer(name_prefix ='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
-    @CommunicationResourcePreparer(resource_group_parameter_name='rg')
-    def test_rooms_update_join_policy_after_start(self, communication_resource_info):
-        os.environ['AZURE_COMMUNICATION_CONNECTION_STRING'] = communication_resource_info[1]
-        
-        # create a room first 
-        room = self.cmd('az communication rooms create').get_output_in_json()
-        self.kwargs.update({
-            'room_id': room['id'],
-            'join_policy': 'CommunicationServiceUsers'})
-
-        # update the room join policy
-        with self.assertRaises(Exception) as raises:
-            self.cmd('az communication rooms update --room {room_id} --join-policy {join_policy}')
-
-        assert 'Bad Request' in str(raises.exception.reason)
-        assert raises.exception.status_code == 400
-        
-        err = raises.exception.model.error
-        assert err.code == 'InvalidInput'
-        assert 'Room join policy cannot be updated after room start time has elapsed' in err.message
-        
-
     @ResourceGroupPreparer(name_prefix ='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
     @CommunicationResourcePreparer(resource_group_parameter_name='rg')
     def test_rooms_update_valid_elapsed_time(self, communication_resource_info):
@@ -187,8 +94,8 @@ class CommunicationRoomsScenarios(ScenarioTest):
         room = self.cmd('az communication rooms create').get_output_in_json()
 
         # update the room with valid elapsed time (to do: joanna_jiang updates the validation time range in every 180 dyas) 
-        validFrom = '2023-01-24T23:09:10.357939+00:00'
-        validUntil = '2023-06-25T13:09:10.357939+00:00'
+        validFrom = (datetime.utcnow() + timedelta(minutes=60)).isoformat()
+        validUntil = (datetime.utcnow() + timedelta(minutes=120)).isoformat()
         
         self.kwargs.update({
             'room_id': room['id'],
@@ -219,35 +126,20 @@ class CommunicationRoomsScenarios(ScenarioTest):
             'validFrom': validFrom,
             'validUntil': validUntil})
 
-        with self.assertRaises(SystemExit) as raises:
+        with self.assertRaises(HttpResponseError) as raises:
             self.cmd('az communication rooms update --room {room_id} --valid-from {validFrom} --valid-until {validUntil}')
 
-        assert 'is not a valid ISO-8601 datetime' in str(raises.exception)
+        assert 'The time range end is in the past' in str(raises.exception.message)
 
 
-    @ResourceGroupPreparer(name_prefix ='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
+    @ResourceGroupPreparer(name_prefix='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
     @CommunicationResourcePreparer(resource_group_parameter_name='rg')
-    def test_rooms_update_attendee_participant(self, communication_resource_info):
+    def test_rooms_list(self, communication_resource_info):
         os.environ['AZURE_COMMUNICATION_CONNECTION_STRING'] = communication_resource_info[1]
-   
-        # create a new room and a new participant
-        room = self.cmd('az communication rooms create').get_output_in_json()
-        attendee_participants_id = self.__create_user(communication_resource_info)
-        
-        # update the room with newly created attendee participant
-        self.kwargs.update({
-            'room_id': room['id'],
-            'attendee_participants_id': attendee_participants_id})
-        self.cmd('az communication rooms update --room {room_id} --attendee-participants {attendee_participants_id}')
+        room = self.cmd('az communication rooms list').get_output_in_json()
 
-        updated_room = self.cmd('az communication rooms get --room {room_id}').get_output_in_json()
-        participants = updated_room['participants']
 
-        # verify the participants role and the number of participants 
-        assert len(participants) == 1
-        assert participants[0]['role'] == 'Attendee'
-       
-
+    #rooms participant management test
     @ResourceGroupPreparer(name_prefix ='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
     @CommunicationResourcePreparer(resource_group_parameter_name='rg')
     def test_rooms_get_default_participants(self, communication_resource_info):
@@ -259,12 +151,10 @@ class CommunicationRoomsScenarios(ScenarioTest):
         # get the default null participant from the newly created room 
         self.kwargs.update({
             'room_id': room['id']})
-        self.cmd('az communication rooms participant get --room {room_id}')
         get_participant = self.cmd('az communication rooms participant get --room {room_id}').get_output_in_json()
-        participants = get_participant['participants']
 
         # verify the defualt participants is null 
-        assert len(participants) == 0
+        assert len(get_participant) == 0
 
        
     @ResourceGroupPreparer(name_prefix ='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
@@ -282,12 +172,12 @@ class CommunicationRoomsScenarios(ScenarioTest):
         self.kwargs.update({
             'room_id': room['id'],
             'participants_id': participant_id})
-        self.cmd('az communication rooms participant add --presenter-participants {participants_id} --room {room_id}')
+        self.cmd('az communication rooms participant add-or-update --presenter-participants {participants_id} --room {room_id}')
         
         get_participant = self.cmd('az communication rooms participant get --room {room_id}').get_output_in_json()
         
-        assert len(get_participant['participants']) == 1
-        role = get_participant['participants'][0]['role']
+        assert len(get_participant) == 1
+        role = get_participant[0]['role']
 
         # verify the type of added participant
         assert role == 'Presenter'
@@ -310,45 +200,11 @@ class CommunicationRoomsScenarios(ScenarioTest):
             'participants_id': participant_id})
         
         with self.assertRaises(Exception) as raises:
-            self.cmd('az communication rooms participant add --consumer-participants {participants_id} --room {room_id}')
+            self.cmd('az communication rooms participant add-or-update --consumer-participants {participants_id} --room {room_id}')
 
         assert 'Bad Request' in str(raises.exception.reason)
         assert raises.exception.status_code == 400
-        
-        err = raises.exception.model.error
-        assert err.code == 'InvalidInput'
-        assert 'Identifier format is invalid' in err.message
-
-
-    @ResourceGroupPreparer(name_prefix ='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
-    @CommunicationResourcePreparer(resource_group_parameter_name='rg')
-    def test_rooms_update_participant_with_same_participant_id(self, communication_resource_info):
-        os.environ['AZURE_COMMUNICATION_CONNECTION_STRING'] = communication_resource_info[1]
-
-        # create a new room 
-        room = self.cmd('az communication rooms create').get_output_in_json()
-
-        # create a valid participant
-        participant_id = self.__create_user(communication_resource_info)
-
-        # add valid participant into the created room 
-        self.kwargs.update({
-            'room_id': room['id'],
-            'participants_id': participant_id})
-     
-        # update the type of participants from presenter to attendee with the same participant id 
-        with self.assertRaises(HttpResponseError) as raises:
-            self.cmd('az communication rooms participant update --presenter-participants {participants_id} --attendee-participants {participants_id} --room {room_id}')
-
-        assert 'Bad Request' in str(raises.exception.reason)
-        assert raises.exception.status_code == 400
-        
-        res = raises.exception.model.additional_properties
-        assert res['status'] == 400
-        assert res['traceId'] is not None
-        assert 'validation errors' in str(res['title'])
-        assert len(res['errors']['Participants']) == 1
-        assert 'duplicate participant' in str(res['errors']['Participants'])
+        assert 'Invalid value for the Participants' in str(raises.exception)
 
 
     @ResourceGroupPreparer(name_prefix ='clitestcommunication_MyResourceGroup'[:7], key='rg', parameter_name ='rg')
@@ -369,13 +225,12 @@ class CommunicationRoomsScenarios(ScenarioTest):
             'presenter_id': presenter_id, 
             'consumer_id': consumer_id})
         
-        self.cmd('az communication rooms participant add --presenter-participants {presenter_id} --consumer-participants {consumer_id} --room {room_id}')
+        self.cmd('az communication rooms participant add-or-update --presenter-participants {presenter_id} --consumer-participants {consumer_id} --room {room_id}')
         
         # get updated room properties
-        updated_room = self.cmd('az communication rooms participant get --room {room_id}').get_output_in_json()
+        participants = self.cmd('az communication rooms participant get --room {room_id}').get_output_in_json()
 
         # verify the length of participant
-        participants = updated_room['participants']
         assert len(participants) == 2 
         
         presenter_role = participants[0]['role']
@@ -401,13 +256,12 @@ class CommunicationRoomsScenarios(ScenarioTest):
         self.kwargs.update({
             'room_id': room['id'],
             'participant_id': participant_id})
-        self.cmd('az communication rooms participant add --consumer-participants {participant_id} --room {room_id}')
+        self.cmd('az communication rooms participant add-or-update --consumer-participants {participant_id} --room {room_id}')
        
         # remove the participant from the created room 
         self.cmd('az communication rooms participant remove --participants {participant_id} --room {room_id} --yes')
-        get_participant = self.cmd('az communication rooms participant get --room {room_id}').get_output_in_json()
-        participant = get_participant['participants']
+        participants = self.cmd('az communication rooms participant get --room {room_id}').get_output_in_json()
 
         # verify the participant list is empty
-        assert len(participant) == 0 
+        assert len(participants) == 0